TURMERIC-1656: Updates the Turmeric Service Creation docs.
This tutorial will replace the existing tutorial on creating a service.
It walks the process from creation of the service to testing of the service
using the jetty-maven-plugin.

Documentaton is written using docbook and can be generated to multiple output
formats to support a variety of reading devices.  Initial formats supported
are HTML and EPub.